Finished Log Routes (Last Routes)

+++ server/controllers/category.js
@@ -1,4 +1,6 @@
 import Category from '../models/Category.js';
+import Product from '../models/Product.js';
+import { createLogMiddleware } from './log.js';
 
 /* CREATE CATEGORY */
 export const createCategory = async (req, res) => {
@@ -29,8 +31,17 @@ export const createCategory = async (req, res) => {
                 res.status(401).json({ error: "Category Already Exists" });
             } else {
                 const savedCategory = await newCategory.save()
+
+                /* LOG PARAMETERS */
+                req.body.info = savedCategory;
+                req.body.type = "create-category";
+
                 console.log(savedCategory);
                 res.status(201).json({ category: savedCategory });
+
+                setTimeout(() => {
+                    createLogMiddleware(req);
+                }, 0);
             }
         }
 
@@ -95,7 +106,15 @@ export const updateCategory = async (req, res) => {
                 if (!result) {
                     return res.status(404).json({ error: "Category Not Found" })
                 } else {
-                    return res.status(200).send({ category: result })
+                    /* LOG PARAMETERS */
+                    req.body.info = result;
+                    req.body.type = "update-category";
+
+                    res.status(200).send({ category: result })
+
+                    setTimeout(() => {
+                        createLogMiddleware(req);
+                    }, 0)
                 }
             }
         }
@@ -110,6 +129,17 @@ export const deleteManyCategories = async (req, res) => {
     try {
         const { idsToDelete } = req.query;
 
+        const categoriesToDelete = await Category.find({ _id: { $in: idsToDelete } })
+
+        for (let i = 0; i < categoriesToDelete.length; i++) {
+            const itemDependent = await Product.findOne({ productCategory: categoriesToDelete[i]._id });
+
+            if (itemDependent) {
+                return res.status(404).json({ error: `The category ${categoriesToDelete[i].categoryName} cannot be deleted because it is linked to other records.` });
+            }
+
+        }
+
         const result = await Category.deleteMany({
             _id: { $in: idsToDelete }
         });
@@ -117,7 +147,15 @@ export const deleteManyCategories = async (req, res) => {
         if (!result) {
             return res.status(404).json({ error: "Categories not Found" });
         } else {
-            return res.status(200).json({ msg: "Categories Successfully Deleted" });
+            /* LOG PARAMETERS */
+            req.body.info = categoriesToDelete;
+            req.body.type = "delete-many-categories";
+
+            res.status(200).json({ msg: "Categories Successfully Deleted" });
+
+            setTimeout(() => {
+                createLogMiddleware(req);
+            }, 0)
         }
 
     } catch (error) {
@@ -130,12 +168,29 @@ export const deleteManyCategories = async (req, res) => {
 export const deleteCategory = async (req, res) => {
     try {
         const { id } = req.params;
+
+        const categoryToDelete = await Category.findById(id)
+
+        const itemDependent = await Product.findOne({ productCategory: categoryToDelete._id });
+
+        if (itemDependent) {
+            return res.status(404).json({ error: `The category ${categoryToDelete.categoryName} cannot be deleted because it is linked to other records.` });
+        }
+
         const result = await Category.findByIdAndDelete(id);
 
         if (!result) {
             return res.status(404).json({ error: "Category not Found" });
         } else {
-            return res.status(200).json({ msg: "Category Successfully Deleted" });
+            /* LOG PARAMETERS */
+            req.body.info = categoryToDelete;
+            req.body.type = "delete-category";
+
+            res.status(200).json({ msg: "Category Successfully Deleted" });
+
+            setTimeout(() => {
+                createLogMiddleware(req);
+            }, 0)
         }
 
     } catch (error) {
